Add generic lenses
This should resolve #2
