+++ uri-bytestring.cabal
@@ -25,12 +25,17 @@ flag lib-Werror
 library
   exposed-modules:
     URI.ByteString
+    URI.ByteString.Lens
+  other-modules:
+    URI.ByteString.Types
+    URI.ByteString.Internal
 
   build-depends:
 
       attoparsec       >= 0.10    && < 0.13
     , base             >= 4.6     && < 4.8
     , bytestring       >= 0.10.2.0 && < 0.11
+    , profunctors      >= 4.0 && < 5.0
 
   hs-source-dirs:      src
   default-language:    Haskell2010
@@ -55,6 +60,8 @@ test-suite test
     , attoparsec
     , base
     , bytestring
+    , lens
+    , quickcheck-instances
   default-language:    Haskell2010
 
 benchmark bench
